requirements.txt
asn1crypto==0.24.0
atomicwrites==1.1.5
attrs==18.1.0
cachetools==3.0.0
certifi==2017.7.27.1
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
cryptography==2.4.2
cycler==0.10.0
defusedxml==0.5.0
docopt==0.6.2
future==0.17.1
google-api-python-client==1.7.8
google-auth==1.6.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
gspread==3.0.1
httplib2==0.11.3
hurry==1.0
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==2.2.2
more-itertools==4.2.0
nose==1.3.7
oauth2client==4.1.2
oauthlib==2.0.4
pdfkit==0.6.1
pluggy==0.6.0
psycopg2==2.7.6.1
py==1.5.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.19
pydotplus==2.0.2
pyparsing==2.2.0
pytest==3.6.0
python-dateutil==2.7.3
python-social-auth==0.3.6
python-telegram-bot==11.1.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-core==1.4.0
sompy==0.1.1
uritemplate==3.0.0
urllib3==1.22
userprofile==0.2.5
websocket-client==0.47.0
Werkzeug==0.12.2
